chat_gpt/requirements.txt
OpenAIAuth>=0.3.2
PyYAML==6.0
Flask==2.2.3
python-dotenv==0.21.1
revChatGPT>=3.1.5
